Move assistant in services folder
